Numerical analysis Homework
These were the assignments i did for my numerical analysis class.
currently they dont have descriptions on what they do mostly because
that was a written report turned in along side them, however if you have
questions on what is going on in any of them feel free to ask.